Refactor and extract `Input` module
