+++ lib/formulator.ex
@@ -1,6 +1,7 @@
 defmodule Formulator do
   use Phoenix.HTML
   alias Formulator.HtmlError
+  import Formulator.Input
 
   @doc  """
   Returns an html input with an associated label.
@@ -103,21 +104,31 @@ defmodule Formulator do
 
   @spec input(Phoenix.HTML.Form.t, atom, []) :: binary
   def input(form, field, options \\ []) do
-    form
-    |> build_input_and_associated_tags(field, options)
-    |> add_wrapper(options)
-  end
-
-  defp build_input_and_associated_tags(form, field, options) do
     {label_options, options} = extract_label_options(options)
-    case label_options do
-      false -> input_without_label(form, field, options)
-      _ -> input_with_label(form, field, label_options, options)
+
+    input_wrapper form, field, options, label_options, fn error ->
+      build_input(form, field, options, error)
     end
   end
 
-  defp add_wrapper(html, options) do
-    [content_tag(:div, html, class: wrapper_class(options))]
+  defp input_wrapper(form, field, options, label_options, fun) do
+    [
+      content_tag(
+        :div,
+        build_input_and_associated_tags(form, field, label_options, fun),
+        class: wrapper_class(options)
+      )
+    ]
+  end
+
+  defp build_input_and_associated_tags(form, field, label_options, fun) do
+    error = HtmlError.html_error(form, field)
+    [
+      build_label(form, field, label_options),
+      fun.(error),
+      error.html
+    ]
+    |> Enum.reject(&(is_nil(&1)))
   end
 
   defp wrapper_class(options) do
@@ -131,87 +142,7 @@ defmodule Formulator do
     {label_options, options}
   end
 
-  defp input_without_label(form, field, options) do
-    options = options ++ build_aria_label(field)
-    error = HtmlError.html_error(form, field)
-    [
-      build_input(form, field, options, error)
-    ] ++ error.html
-  end
-
-  defp input_with_label(form, field, label_options, options) do
-    error = HtmlError.html_error(form, field)
-    build_html(form, field, label_options, options, error)
-  end
-
-  defp build_aria_label(field) do
-    ["aria-label": format_label(field)]
-  end
-
-  defp format_label(field) do
-    field |> to_string |> String.replace("_", " ") |> String.capitalize
-  end
-
-  defp build_html(form, field, label_options, options, error) do
-    [
-      build_label(form, field, label_options),
-      build_input(form, field, options, error)
-    ] ++ error.html
-  end
-
-  defp build_input(form, field, options, error) do
-    input_type = options[:as] || :text
-    input_class = options[:class] || ""
-
-    options =
-      options
-      |> add_validation_attributes(form, field)
-      |> add_format_validation_attribute(form, field)
-      |> Keyword.delete(:as)
-      |> Keyword.put(:class, add_error_class(input_class, error.class))
-
-    apply(Phoenix.HTML.Form, input_function(input_type), [form, field, options])
-  end
-
-  defp add_validation_attributes(options, %{impl: impl, source: %{validations: _}} = form, field) when is_atom(impl) do
-    if option_enabled?(options, :validate, true) do
-      form
-      |> Phoenix.HTML.Form.input_validations(field)
-      |> Keyword.merge(options)
-      |> Keyword.delete(:validate)
-    else
-      options
-    end
-  end
-  defp add_validation_attributes(options, _, _), do: options
-
-  defp add_format_validation_attribute(options, %{impl: impl, source: %{validations: _}} = form, field) when is_atom(impl) do
-    with true <- option_enabled?(options, :validate_regex, true),
-      {:format, regex} <- form.source.validations[field]
-    do
-      options
-      |> Keyword.put_new(:pattern, Regex.source(regex))
-      |> Keyword.delete(:validate_regex)
-    else
-      _ -> options
-    end
-  end
-  defp add_format_validation_attribute(options, _, _), do: options
-
-  defp option_enabled?(options, field, default) do
-    Enum.any?([
-      options[field] == true,
-      Application.get_env(:formulator, field, default) == true,
-    ])
-  end
-
-  defp add_error_class(input_class, error_class) do
-    [input_class, error_class]
-    |> Enum.reject(&(is_nil(&1)))
-    |> Enum.join(" ")
-    |> String.trim
-  end
-
+  def build_label(_form, _field, false), do: nil
   def build_label(form, field, label_text) when is_binary(label_text) do
     build_label(form, field, [text: label_text])
   end
@@ -222,11 +153,4 @@ defmodule Formulator do
       text -> label(form, field, text, label_options |> Keyword.delete(:text))
     end
   end
-
-  defp input_function(:checkbox), do: :checkbox
-  defp input_function(:date), do: :date_select
-  defp input_function(:datetime), do: :datetime_select
-  defp input_function(:time), do: :time_select
-  defp input_function(:textarea), do: :textarea
-  defp input_function(input_type), do: :"#{input_type}_input"
 end
